requirements.txt
pyyaml==5.3.1
python-designateclient==4.1.0
flake8==3.8.4
